# swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/notificacion/{_id}": {
            "get": {
                "operationId": "NotificacionController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "notificacion"
                ]
            },
            "put": {
                "operationId": "NotificacionController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "notificacion"
                ]
            },
            "delete": {
                "operationId": "NotificacionController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "notificacion"
                ]
            }
        },
        "/plantilla": {
            "post": {
                "operationId": "PlantillaController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlantillaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "plantilla"
                ]
            },
            "get": {
                "operationId": "PlantillaController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plantilla"
                ]
            }
        },
        "/plantilla/{_id}": {
            "get": {
                "operationId": "PlantillaController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plantilla"
                ]
            },
            "put": {
                "operationId": "PlantillaController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlantillaDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plantilla"
                ]
            },
            "delete": {
                "operationId": "PlantillaController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "plantilla"
                ]
            }
        },
...
        "/tipo_notificacion": {
            "post": {
                "operationId": "TipoNotificacionController_post",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoNotificacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo_notificacion"
                ]
            },
            "get": {
                "operationId": "TipoNotificacionController_getAll",
                "parameters": [
                    {
                        "name": "query",
                        "required": false,
                        "in": "query",
                        "description": "query - Filter. e.g. col1:v1,col2:v2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fields",
                        "required": false,
                        "in": "query",
                        "description": "fields - Fields returned. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortby",
                        "required": false,
                        "in": "query",
                        "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "required": false,
                        "in": "query",
                        "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": false,
                        "in": "query",
                        "description": "limit - Limit the size of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "required": false,
                        "in": "query",
                        "description": "offset - Start position of result set. Must be an integer",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "populate",
                        "required": false,
                        "in": "query",
                        "description": "populate - show subqueries. Must be a boolean",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo_notificacion"
                ]
            }
        },
        "/tipo_notificacion/{_id}": {
            "get": {
                "operationId": "TipoNotificacionController_getById",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo_notificacion"
                ]
            },
            "put": {
                "operationId": "TipoNotificacionController_put",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoNotificacionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo_notificacion"
                ]
            },
            "delete": {
                "operationId": "TipoNotificacionController_delete",
                "parameters": [
                    {
                        "name": "_id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "tipo_notificacion"
                ]
            }
        }
    },
    "info": {
        "title": "notificaciones_crud",
        "description": "API CRUD sistema de notificaciones",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "NotificacionDto": {
                "type": "object",
                "properties": {
                    "sistema_id": {
                        "type": "string"
                    },
                    "tipo_notificacion_id": {
                        "type": "string"
                    },
                    "destinatario": {
                        "type": "string"
                    },
                    "destinatarios": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "remitente": {
                        "type": "string"
                    },
                    "asunto": {
                        "type": "string"
                    },
                    "mensaje": {
                        "type": "string"
                    },
                    "lectura": {
                        "type": "boolean",
                        "default": false
                    },
                    "fecha_lectura": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "metadatos": {
                        "type": "object"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "sistema_id",
                    "tipo_notificacion_id",
                    "destinatario",
                    "destinatarios",
                    "remitente",
                    "asunto",
                    "mensaje",
                    "lectura",
                    "fecha_lectura",
                    "metadatos",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
            "PlantillaDto": {
                "type": "object",
                "properties": {
                    "sistema_id": {
                        "type": "string"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "codigo_abreviacion": {
                        "type": "string"
                    },
                    "plantilla_mensaje": {
                        "type": "string"
                    },
                    "metadatos": {
                        "type": "object"
                    },
                    "activo": {
                        "type": "boolean"
                    },
                    "fecha_creacion": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "fecha_modificacion": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "sistema_id",
                    "nombre",
                    "descripcion",
                    "codigo_abreviacion",
                    "plantilla_mensaje",
                    "metadatos",
                    "activo",
                    "fecha_creacion",
                    "fecha_modificacion"
                ]
            },
...
            "TipoNotificacionDto": {
                "type": "object",
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "codigo_abreviacion": {
                        "type": "string"
                    }
                },
                "required": [
                    "nombre",
                    "descripcion",
                    "codigo_abreviacion"
                ]
            }
        }
    }
}